# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/fpt/team5/golddigger/dal/MyDbContext.java

+++ app/src/main/java/com/fpt/team5/golddigger/BalanceActivity.java
@@ -58,7 +58,7 @@ public class BalanceActivity extends AppCompatActivity {
     }
 
     private void OnBtnStartClick(View view) {
-        float amount = Float.parseFloat(balanceEt.getText().toString());
+        double amount = Double.parseDouble(balanceEt.getText().toString());
         String title = titleEt.getText().toString();
